ipq807x: CIG wf660ag: support both EMMC and SPI NOR+EMCC H/W
Using one firmware to support the board which has EMMC only and SPI NOR+EMMC equipped

Fixes: WIFI-12941

+++ feeds/ipq807x/ipq807x/base-files/lib/upgrade/platform.sh
@@ -52,7 +52,7 @@ do_flash_emmc() {
 	tar Oxf $tar_file ${board_dir}/$part | dd of=${emmcblock}
 }
 
-emmc_do_upgrade_bootconfig() {
+spi_nor_emmc_do_upgrade_bootconfig() {
 	local tar_file="$1"
 
 	local board_dir=$(tar tf $tar_file | grep -m 1 '^sysupgrade-.*/$')
@@ -84,11 +84,17 @@ emmc_do_upgrade_bootconfig() {
 	fi
 
 	for part in "0:BOOTCONFIG" "0:BOOTCONFIG1"; do
-		local emmcblock=$(find_mmc_part $part)
-		echo erase ${emmcblock}
-		dd if=/dev/zero of=${emmcblock} 2> /dev/null
-		echo update $emmcblock
-		dd if=/tmp/bootconfig of=${emmcblock} 2> /dev/null
+               local mtdchar=$(echo $(find_mtd_chardev $part) | sed 's/^.\{5\}//')
+               if [ -n "$mtdchar" ]; then
+                       echo start to update $mtdchar
+                       mtd -qq write /proc/boot_info/getbinary_bootconfig "/dev/${mtdchar}" 2>/dev/null && echo update mtd $mtdchar
+               else
+                       emmcblock=$(find_mmc_part $part)
+                       echo erase ${emmcblock}
+                       dd if=/dev/zero of=${emmcblock} 2> /dev/null
+                       echo update $emmcblock
+                       dd if=/tmp/bootconfig of=${emmcblock} 2> /dev/null
+               fi
 	done
 }
 
@@ -175,7 +181,7 @@ platform_do_upgrade() {
 		qca_do_upgrade $1
 		;;
 	cig,wf660a)
-		emmc_do_upgrade_bootconfig $1
+		spi_nor_emmc_do_upgrade_bootconfig $1
 		;;
 	motorola,q14)
 		emmc_do_upgrade $1
